documenting the data generation and summary stats calculation

+++ flares/flares.py
@@ -7,6 +7,8 @@ MIT License (2022)
 """
 
 import warnings
+from os.path import exists
+
 from datetime import datetime
 
 import numpy as np
@@ -142,8 +144,14 @@ def get_flares(u_ld, flc, emin, emax, errval, spot_radius, n_inclinations,
     
     # write results to file if any flares were found 
     if flares.shape[0] > 0:
-        with open(path, "a") as file:
-            flares.to_csv(file, index=False, header=False)
+        # write header if necessary, but only once
+    	file_exists = exists(path)
+	    if file_exists:
+	        with open(path, "a") as file:
+	            flares.to_csv(file, index=False, header=False)
+	    else:
+	        with open(path, "a") as file:
+	            flares.to_csv(file, index=False, header=True)
 
     return flares
